Fix bug & rename auto file upload component
